Update custom image to use Next Image component for optimization
